Continue simplifying db stuff

+++ app/lib/user_token_struct_utils.py
@@ -30,6 +30,6 @@ class UserTokenStructUtils:
 
 @cython.cfunc
 def _add_b32_padding(s: str) -> str:
-    s_len: cython.int = len(s)
+    s_len: cython.Py_ssize_t = len(s)
     pad_len: cython.int = int(ceil(s_len / 8) * 8 - s_len)
     return s + ('=' * pad_len)
